// src/api/enums/property-primitives.enum.ts
/*
 * This code was generated automatically
 * and should not be modifiyed manyally,
 * becouse it can be overwritten in any
 * moment. All modifications are allowed
 * in nested class file
 * in a proper way.
 */

export enum PropertyPrimitivesEnum{
	Int = 1,
	Decimal = 2,
	ShortText = 3,
	LongText = 4,
	IntWithUnit = 5,
	DecimalWithUnit = 6,
	File = 7,
	Boolean = 8,
	Image = 9,
}